+++ shared/components/src/resolver.ts
@@ -1,31 +1,31 @@
-import type { ComponentResolver } from 'unplugin-vue-components';
+import type { ComponentResolver } from "unplugin-vue-components";
 
 const components = [
-  'HButton',
-  'HCardTitle',
-  'HCenterLabel',
-  'HChartCard',
-  'HColumn',
-  'HContainer',
-  'HDateTime',
-  'HDivider',
-  'HDialog',
-  'HDuration',
-  'HIconButton',
-  'HIconSelect',
-  'HFieldLabel',
-  'HTreeField',
-  'HLabel',
-  'HListItem',
-  'HLoading',
-  'HPagination',
-  'HParticles',
-  'HRow',
-  'HSelect',
-  'HSignInBackground',
-  'HSwitch',
-  'HTextField',
-  'HVisibilityButton',
+  "HButton",
+  "HCardTitle",
+  "HCenterLabel",
+  "HChartCard",
+  "HColumn",
+  "HContainer",
+  "HDateTime",
+  "HDivider",
+  "HDialog",
+  "HDuration",
+  "HIconButton",
+  "HIconSelect",
+  "HFieldLabel",
+  "HTreeField",
+  "HLabel",
+  "HListItem",
+  "HLoading",
+  "HPagination",
+  "HParticles",
+  "HRow",
+  "HSelect",
+  "HSignInBackground",
+  "HSwitch",
+  "HTextField",
+  "HVisibilityButton",
 ];
 
 function isValidComponent(name: string) {
@@ -36,15 +36,15 @@ function isValidComponent(name: string) {
  */
 export function HerodotusResolver(): ComponentResolver {
   return {
-    type: 'component',
+    type: "component",
     resolve: (name: string) => {
       // 只处理指定前缀的组件
-      if (name.startsWith('H') && isValidComponent(name)) {
+      if (name.startsWith("H") && isValidComponent(name)) {
         const componentName = name;
-        const from = '@herodotus-cloud/components';
+        const from = "@herodotus-cloud/components";
 
         // 构建 side effects 配置
-        const sideEffects: string[] = ['@herodotus-cloud/components/style.css'];
+        const sideEffects: string[] = ["@herodotus-cloud/components/style.css"];
 
         return {
           name: componentName,